Removed Flagging/Updated Style
