+++ auth-web/src/components/auth/TestimonialQuotes.vue
@@ -1,7 +1,7 @@
 <template>
   <section id="testimonials-container">
     <v-container class="quote-container">
-      <div class="fading">
+      <div>
         <p class="quote-text">{{quotes[quoteIndex].text}}</p>
         <p class="quote-author">&ndash; {{quotes[quoteIndex].author}}</p>
       </div>
